+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/guide/BookSlimefunGuide.java
@@ -210,7 +210,7 @@ public class BookSlimefunGuide implements SlimefunGuideImplementation {
     private void addSlimefunItem(Category category, int page, Player p, PlayerProfile profile, SlimefunItem item, List<ChatComponent> items) {
         NamespacedKey key = new NamespacedKey(SlimefunPlugin.instance(), item.getId().toLowerCase(Locale.ROOT));
 
-        if (!Slimefun.hasUnlocked(p, item, false) && item.getResearch() != null) {
+        if (!item.canUse(p, false) && item.hasResearch()) {
             Research research = item.getResearch();
 
             ChatComponent component = new ChatComponent(ChatUtils.crop(ChatColor.RED, item.getItemName()) + "\n");