+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/listeners/BackpackListener.java
@@ -7,6 +7,7 @@ import java.util.UUID;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import javax.annotation.ParametersAreNonnullByDefault;
 
 import org.apache.commons.lang.Validate;
 import org.bukkit.ChatColor;
@@ -30,7 +31,6 @@ import io.github.thebusybiscuit.slimefun4.implementation.SlimefunPlugin;
 import io.github.thebusybiscuit.slimefun4.implementation.items.backpacks.Cooler;
 import io.github.thebusybiscuit.slimefun4.implementation.items.backpacks.SlimefunBackpack;
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
-import me.mrCookieSlime.Slimefun.api.Slimefun;
 
 /**
  * This {@link Listener} is responsible for all events centered around a {@link SlimefunBackpack}.
@@ -57,12 +57,21 @@ public class BackpackListener implements Listener {
 
     @EventHandler
     public void onClose(InventoryCloseEvent e) {
-        Player p = ((Player) e.getPlayer());
+        Player p = (Player) e.getPlayer();
+
+        if (markBackpackDirty(p)) {
+            p.playSound(p.getLocation(), Sound.ENTITY_HORSE_ARMOR, 1F, 1F);
+        }
+    }
+
+    private boolean markBackpackDirty(@Nonnull Player p) {
         ItemStack backpack = backpacks.remove(p.getUniqueId());
 
         if (backpack != null) {
-            p.playSound(p.getLocation(), Sound.ENTITY_HORSE_ARMOR, 1F, 1F);
             PlayerProfile.getBackpack(backpack, PlayerBackpack::markDirty);
+            return true;
+        } else {
+            return false;
         }
     }
 
@@ -111,7 +120,7 @@ public class BackpackListener implements Listener {
 
     public void openBackpack(Player p, ItemStack item, SlimefunBackpack backpack) {
         if (item.getAmount() == 1) {
-            if (Slimefun.hasUnlocked(p, backpack, true) && !PlayerProfile.get(p, profile -> openBackpack(p, item, profile, backpack.getSize()))) {
+            if (backpack.canUse(p, true) && !PlayerProfile.get(p, profile -> openBackpack(p, item, profile, backpack.getSize()))) {
                 SlimefunPlugin.getLocalization().sendMessage(p, "messages.opening-backpack");
             }
         } else {
@@ -119,8 +128,10 @@ public class BackpackListener implements Listener {
         }
     }
 
+    @ParametersAreNonnullByDefault
     private void openBackpack(Player p, ItemStack item, PlayerProfile profile, int size) {
         List<String> lore = item.getItemMeta().getLore();
+
         for (int line = 0; line < lore.size(); line++) {
             if (lore.get(line).equals(ChatColor.GRAY + "ID: <ID>")) {
                 setBackpackId(p, item, line, profile.createBackpack(size).getId());
@@ -128,6 +139,15 @@ public class BackpackListener implements Listener {
             }
         }
 
+        /*
+         * If the current Player is already viewing a backpack (for whatever reason),
+         * terminate that view.
+         */
+        if (markBackpackDirty(p)) {
+            p.closeInventory();
+        }
+
+        // Check if someone else is currently viewing this backpack
         if (!backpacks.containsValue(item)) {
             p.playSound(p.getLocation(), Sound.ENTITY_HORSE_ARMOR, 1F, 1F);
             backpacks.put(p.getUniqueId(), item);