+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/listeners/SlimefunItemInteractListener.java
@@ -25,7 +25,6 @@ import io.github.thebusybiscuit.slimefun4.implementation.SlimefunPlugin;
 import io.github.thebusybiscuit.slimefun4.utils.SlimefunUtils;
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
 import me.mrCookieSlime.Slimefun.api.BlockStorage;
-import me.mrCookieSlime.Slimefun.api.Slimefun;
 import me.mrCookieSlime.Slimefun.api.inventory.BlockMenu;
 import me.mrCookieSlime.Slimefun.api.inventory.BlockMenuPreset;
 import me.mrCookieSlime.Slimefun.api.inventory.UniversalBlockMenu;
@@ -94,8 +93,10 @@ public class SlimefunItemInteractListener implements Listener {
         Optional<SlimefunItem> optional = event.getSlimefunItem();
 
         if (optional.isPresent()) {
-            if (Slimefun.hasUnlocked(e.getPlayer(), optional.get(), true)) {
-                return optional.get().callItemHandler(ItemUseHandler.class, handler -> handler.onRightClick(event));
+            SlimefunItem sfItem = optional.get();
+
+            if (sfItem.canUse(e.getPlayer(), true)) {
+                return sfItem.callItemHandler(ItemUseHandler.class, handler -> handler.onRightClick(event));
             } else {
                 event.setUseItem(Result.DENY);
             }
@@ -109,12 +110,14 @@ public class SlimefunItemInteractListener implements Listener {
         Optional<SlimefunItem> optional = event.getSlimefunBlock();
 
         if (optional.isPresent()) {
-            if (!Slimefun.hasUnlocked(event.getPlayer(), optional.get(), true)) {
+            SlimefunItem sfItem = optional.get();
+
+            if (!sfItem.canUse(event.getPlayer(), true)) {
                 event.getInteractEvent().setCancelled(true);
                 return false;
             }
 
-            boolean interactable = optional.get().callItemHandler(BlockUseHandler.class, handler -> handler.onRightClick(event));
+            boolean interactable = sfItem.callItemHandler(BlockUseHandler.class, handler -> handler.onRightClick(event));
 
             if (!interactable) {
                 String id = optional.get().getId();